refactor(firestore): lib/firestore.tsを機能別に4ファイルへ分割
- common.ts: DB接続、ユーティリティ、データ正規化
- userData.ts: 書き込みキュー、ユーザーデータCRUD、リアルタイム購読
- defectBeans.ts: 欠点豆マスタ・ユーザー固有データ操作
- workProgress.ts: 作業進捗の追加・更新・削除・アーカイブ
- index.ts: barrel re-exportで既存インポートパスを維持

Closes #35
